requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic-settings==2.1.0
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.1.2
pyarrow==14.0.1
